+++ src/pages/Home/components/Education/education.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import DegreeDetails from './degreeDetails';
 import { WideCardSection } from '../../common';
-import websiteContent from '../../websiteContent.json';
+import websiteContent from '../../data/websiteContent.json';
 
 const Education = () => {
   const { content, type, sectionTitle } = websiteContent.data.education;