+++ src/pages/MothersDay2021/components/index.js
@@ -1,0 +1,3 @@
+export { default as Envelope } from './Envelope';
+export { default as Mail } from './Mail';
+export { default as MailContent } from './MailContent';