+++ src/pages/MothersDay2021/utils/styles.js
@@ -1,0 +1,9 @@
+export const colors = {
+  blue: '#60a7b9',
+  darkPink: '#e5acb6',
+  lightBlue: '#c9ebe4',
+  lightPink: '#ffd2da',
+  lightYellow: '#ffebb2',
+  pink: '#ffc0cb',
+  white: '#ffffff',
+};